+++ spec/transpec/converter_spec.rb
@@ -5,7 +5,8 @@ require 'transpec/converter'
 
 module Transpec
   describe Converter do
-    subject(:converter) { Converter.new(configuration) }
+    subject(:converter) { Converter.new(configuration, rspec_version) }
+    let(:rspec_version) { Transpec.current_rspec_version }
     let(:configuration) { Configuration.new }
 
     describe '#convert_file!' do
@@ -380,8 +381,8 @@ module Transpec
       let(:method_stub_object) { double('method_stub_object').as_null_object }
 
       shared_examples 'invokes MethodStub#allowize!' do
-        it 'invokes MethodStub#allowize!' do
-          method_stub_object.should_receive(:allowize!)
+        it 'invokes MethodStub#allowize! with #receive_messages availability' do
+          method_stub_object.should_receive(:allowize!).with(rspec_version.receive_messages_available?)
           converter.process_method_stub(method_stub_object)
         end
       end
@@ -484,6 +485,69 @@ module Transpec
       end
     end
 
+    describe '#process_be_boolean' do
+      let(:be_boolean_object) { double('be_boolean_object').as_null_object }
+
+      context 'when RSpecVersion#be_truthy_available? returns true' do
+        before { rspec_version.stub(:be_truthy_available?).and_return(true) }
+
+        context 'when Configuration#convert_deprecated_method? is true' do
+          before { configuration.convert_deprecated_method = true }
+
+          context 'and Configuration#boolean_matcher_type is :conditional' do
+            before { configuration.boolean_matcher_type = :conditional }
+
+            context 'and Configuration#form_of_be_falsey is "be_falsey"' do
+              before { configuration.form_of_be_falsey = 'be_falsey' }
+
+              it 'invokes BeBoolean#convert_to_conditional_matcher! with "be_falsey"' do
+                be_boolean_object.should_receive(:convert_to_conditional_matcher!).with('be_falsey')
+                converter.process_be_boolean(be_boolean_object)
+              end
+            end
+
+            context 'and Configuration#form_of_be_falsey is "be_falsy"' do
+              before { configuration.form_of_be_falsey = 'be_falsy' }
+
+              it 'invokes BeBoolean#convert_to_conditional_matcher! with "be_falsy"' do
+                be_boolean_object.should_receive(:convert_to_conditional_matcher!).with('be_falsy')
+                converter.process_be_boolean(be_boolean_object)
+              end
+            end
+          end
+
+          context 'and Configuration#boolean_matcher_type is :exact' do
+            before { configuration.boolean_matcher_type = :exact }
+
+            it 'invokes BeBoolean#convert_to_exact_matcher!' do
+              be_boolean_object.should_receive(:convert_to_exact_matcher!)
+              converter.process_be_boolean(be_boolean_object)
+            end
+          end
+        end
+
+        context 'when Configuration#convert_deprecated_method? is false' do
+          before { configuration.convert_deprecated_method = false }
+
+          it 'does nothing' do
+            be_boolean_object.should_not_receive(:convert_to_conditional_matcher!)
+            be_boolean_object.should_not_receive(:convert_to_exact_matcher!)
+            converter.process_be_boolean(be_boolean_object)
+          end
+        end
+      end
+
+      context 'when RSpecVersion#be_truthy_available? returns true' do
+        before { rspec_version.stub(:be_truthy_available?).and_return(false) }
+
+        it 'does nothing' do
+          be_boolean_object.should_not_receive(:convert_to_conditional_matcher!)
+          be_boolean_object.should_not_receive(:convert_to_exact_matcher!)
+          converter.process_be_boolean(be_boolean_object)
+        end
+      end
+    end
+
     describe '#process_be_close' do
       let(:be_close_object) { double('be_close_object').as_null_object }
 
@@ -496,7 +560,7 @@ module Transpec
         end
       end
 
-      context 'when Configuration#convert_deprecated_method? is true' do
+      context 'when Configuration#convert_deprecated_method? is false' do
         before { configuration.convert_deprecated_method = false }
 
         it 'does not invoke BeClose#convert_to_be_within!' do
@@ -518,7 +582,7 @@ module Transpec
         end
       end
 
-      context 'when Configuration#convert_deprecated_method? is true' do
+      context 'when Configuration#convert_deprecated_method? is false' do
         before { configuration.convert_deprecated_method = false }
 
         it 'does not invoke BeClose#convert_to_be_within!' do
@@ -540,7 +604,7 @@ module Transpec
         end
       end
 
-      context 'when Configuration#convert_its? is true' do
+      context 'when Configuration#convert_its? is false' do
         before { configuration.convert_its = false }
 
         it 'does not invoke Its#convert_to_describe_subject_it!' do